Refactoring case and adding all scores checking
All available scores are now checked to be non-zero. If any are zero,
the check fill fail adding a MISSING_SCORES issue listing the maximum
scores.

Refactoring of case library.
